ENH: Add ability to choose integration method
One can choose between a "Recursive" integration strategy and a
convolutional one. Both seem to have more or less equal performance on
noisy phantoms, but the recursive strategy has PERFECT peformance on a
noiseless phantom. This makes it preferable. It also seems to be faster
than the convolutional method, at least when tested on Windows.